Copy pom body from existing cinema project
This adds the dependencies, plugins and build details from the "old"
dodgy cinema repository, which is originally taken from John at QA's
example cinema app repo.

It adds managed dependencies, in a particular order, and should suffice
for all basic functionality.

While it is annoying to add all these dependencies at once, upfront,
this does seem to be the easiest way to get the project working so that
development can continue.